src/bot/requirements.txt
telegram
boto3
cv
numpy
requests
pillow
opencv-python
